moved ajax get request to home component instead of carousel component
